Make <code> tags inline and remove wrapper <p>
Two changes: instead of wrapping the `yarn why` message in `<p>` tags, which the browser will remove during parsing since they're around an `<ul>`, I just made the `<ul>` the top-level element.

The other change is that I changed the `<code>` tag from wrapping a block-level element to wrapping each list item; since the browser "fixes" the wrapping of a block-level tag by an inline tag during parsing, it wound up creating an empty `<code>` tag above the list, which bugged me. This way, it's just the contents of each `<li>` that are wrapped, as intended.